+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 <!-- There is always Unreleased section on the top. Subsections (Add, Changed, Fix, Removed) should be Add as needed. -->
 ## Unreleased
+- [**BC**] Add MetaData types
+    - `MetaDataOnlyCreatedAt`
+    - `MetaDataCreatedAndProcessed`
 
 ## 17.2.0 - 2021-10-21
 - Log debug of commit message key, if manual committing is set up